requirements.txt
pyngrok
lime==0.2.0.1
shap==0.44.1
openai==1.12.0
python-dotenv